# src/main/java/com/wallet/application/dto/CreateWalletRequest.java
package com.wallet.application.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Pattern;

public class CreateWalletRequest {
    
    @NotBlank(message = "User ID is required")
    private String userId;
    
    @NotBlank(message = "Currency is required")
    @Pattern(regexp = "^[A-Z]{3}$", message = "Currency must be a 3-letter code (e.g., BRL, USD)")
    private String currency;
    
    public CreateWalletRequest() {
        // Para deserialização JSON
    }
    
    public CreateWalletRequest(String userId, String currency) {
        this.userId = userId;
        this.currency = currency;
    }
    
    public String getUserId() {
        return userId;
    }
    
    public void setUserId(String userId) {
        this.userId = userId;
    }
    
    public String getCurrency() {
        return currency;
    }
    
    public void setCurrency(String currency) {
        this.currency = currency;
    }
} 