# src/main/java/com/wallet/application/dto/TransactionListResponse.java
package com.wallet.application.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class TransactionListResponse {
    private List<TransactionResponse> transactions;
    private int page;
    private int size;
    private long totalElements;
    private int totalPages;
} 